feat: move s3_prune_backups to top level script
